refactor file names and add recycler card view
