+++ app/src/main/java/com/example/mixtape/ui/past/PastViewModel.java
@@ -1,14 +1,14 @@
-package com.example.mixtape.ui.dashboard;
+package com.example.mixtape.ui.past;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class DashboardViewModel extends ViewModel {
+public class PastViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public DashboardViewModel() {
+    public PastViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is dashboard fragment");
     }
