+++ app/src/main/java/com/example/mixtape/ui/profile/ProfileViewModel.java
@@ -1,14 +1,14 @@
-package com.example.mixtape.ui.notifications;
+package com.example.mixtape.ui.profile;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class NotificationsViewModel extends ViewModel {
+public class ProfileViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public NotificationsViewModel() {
+    public ProfileViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is notifications fragment");
     }
