+++ app/src/main/res/drawable/custom_edittext.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="rectangle">
+
+    <stroke
+        android:width="5dp"
+        android:color="@color/black" />
+
+    <corners
+        android:radius="30dp"
+        />
+
+</shape>
